requirements.txt
tqdm
pytorch
torchvision
pyyaml
tensorboardX
scikit-learn
matplotlib
